requirements.txt
elasticsearch==2.3.0
weibo==0.2.2
django-dotenv==1.3.0
django-getenv==1.3.1
